perf: optimize katgpt-core — layout, SIMD, allocations, error handling
- Layout: reorder struct fields in 20+ structs to eliminate padding (u64→u32→u8)
- Layout: u8→u32 for TripleEvidence.count and ReconstructionResult.steps to remove trailing padding
- Layout: remove unnecessary #[repr(C)] from SpectralThreatFeatures (all-f32, no padding difference)
- Enum: add #[repr(u8)] to SubStepStrategy and DecBackend (field-less, 1-byte discriminant)
- SIMD: merge double-pass loops in compute_retrieval_margin() and maxsim_score_packed()
- SIMD: branch-free sign multiply + 4-wide chunked loops in dec/operators codifferential/graph_laplacian
- SIMD: eliminate double cast in sigmoid_margin_loss (u32→f32 instead of u32→u32→f32)
- Errors: replace let _ = core::fmt::write() with explicit is_err() in questbench
- Errors: replace let _ = select_nth_unstable with tuple destructuring in top_k
- Errors: prefix unused rank param in hodge_star with _rank

+++ crates/katgpt-core/src/sense/brain.rs
@@ -349,7 +349,7 @@ impl NpcBrain {
         ReconstructionResult {
             passive,
             active,
-            steps: state.step(),
+            steps: state.step() as u32,
             evidence: state.evidence().clone(),
             hla_delta,
         }
